+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 7. The card "4" has no special effect, the card "2" has it - fixed.
 8. Missing brackets - fixed.
 9. King changing direction of the game - fixed
+10. Player can only discard the same value or suit card or Ace or Queen - fixed.
 
 * v1.1.0 [2019-11-08]: Added a SmartAI computer opponent.
   Added strategy players.SmartAI